cn_server/countrytotaltests_service/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4
pyspark ~= 3.1.1